Fix concerning _perceptron.
Had an issue with using

this->_perceptron.reset(new Perceptron(rows, cols, activation))

(did not work in this scope). Instead we followed the compiler's advice

this->_perceptron.reset(new class Perceptron(rows, cols, activation))

Not quite clear why the above did not work. When googling the error we
only found issues where e.g. the name of an object was the same as the class...